requirements.txt
python-dotenv==1.0.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
flask==3.0.2
Flask-Cors==4.0.0
greenlet==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
SQLAlchemy==1.3.23
typing-extensions==4.10.0
werkzeug==3.0.2
zipp==3.18.1
